added confirm message before deleting the multiple problems
