+++ src/components/ui/multiSelectDropdown.tsx
@@ -5,58 +5,14 @@ import { useState } from "react";
 import { Table } from "@tanstack/react-table";
 import { Checkbox } from "./checkbox";
 import { ProblemRow } from "../DataTableContent";
-import { api } from "~/trpc/react";
-import { useRouter } from "next/navigation";
-import { useToast } from "./use-toast";
 
 const MultiColumnDropdown = ({ table } : { table: Table<ProblemRow> }) => {
   const [isDropdownOpen, setIsDropdownOpen] = useState(false);
 
-  const router = useRouter();
-  const utils = api.useUtils();
-  const { toast } = useToast();
-
-  const deleteSelectedProblems = api.problem.deleteMultipleProblems.useMutation({
-    onSuccess: () => {
-      router.refresh();
-      utils.problem.getAllProblems.invalidate();
-      toast({
-        title: "DELETED SELECTED PROBLEMS SUCCESSFULLY",
-        description: "Problems saved successfully to the database",
-      })
-      setIsDropdownOpen(false);
-      
-    }, 
-    onError: (error) => {
-      const errorMessage = error.data?.zodError?.fieldErrors.content![0];
-      console.log(error, "Error.........");
-      if (errorMessage) {
-        toast({
-          title: "ERROR DELETING SELECTED PROBLEMS",
-          description: errorMessage,
-        })
-      } else {
-        toast({
-          title: "ERROR DELETING",
-          description: "Failed to delete selected problems!",
-        })
-      }
-    }
-  });
-
-
-  const handleDeleteSelected = () => {
-    const selectedRows = Object.keys(table.getState().rowSelection);
-    const rowIds = selectedRows.map(rowId => table.getRow(rowId).original.id);
-    console.log(selectedRows, rowIds)
-
-    deleteSelectedProblems.mutate({ problemIds: rowIds })
-  } 
-
   return (
     <DropdownMenu open={isDropdownOpen} onOpenChange={(value) => setIsDropdownOpen(value)}>
     <DropdownMenuTrigger asChild>
-      <Button variant="ghost" className="h-8 w-20 ml-6">
+      <Button variant="ghost" className="h-8 w-15 py-2 px-0">
         <span className="sr-only">Open menu</span>
         <MoreHorizontal className="h-4 w-4" />
       </Button>
@@ -79,9 +35,6 @@ const MultiColumnDropdown = ({ table } : { table: Table<ProblemRow> }) => {
         </DropdownMenuItem>
         <DropdownMenuItem onClick={() => {}}>
           Edit
-        </DropdownMenuItem>
-        <DropdownMenuItem onClick={handleDeleteSelected}>
-          Delete Selected
         </DropdownMenuItem>
         <DropdownMenuItem>
           Remove
